+++ shaders/Texture.vert
@@ -1,4 +1,4 @@
-#version 130
+#version 120
 uniform int uHasHeightMap;
 uniform sampler2D uHeightMap;
 
@@ -17,16 +17,16 @@ attribute vec3 aTexCoord;
 attribute vec3 aPosition;
 attribute vec3 aNormal;
 
-attribute int aBoneID0;
-attribute int aBoneWeight0;
-attribute int aBoneID1;
-attribute int aBoneWeight1;
-attribute int aBoneID2;
-attribute int aBoneWeight2;
-attribute int aBoneID3;
-attribute int aBoneWeight3;
-attribute int aBoneID4;
-attribute int aBoneWeight4;
+attribute float aBoneID0;
+attribute float aBoneWeight0;
+attribute float aBoneID1;
+attribute float aBoneWeight1;
+attribute float aBoneID2;
+attribute float aBoneWeight2;
+attribute float aBoneID3;
+attribute float aBoneWeight3;
+attribute float aBoneID4;
+attribute float aBoneWeight4;
 
 varying vec4 vPosition;
 varying vec2 vTexCoord;
@@ -39,19 +39,19 @@ void main() {
    mat4 bone = mat4(1.0);
    if (uHasBones != 0) {
       if (aBoneID0 != -1) {
-         bone = uBones[aBoneID0] * aBoneWeight0;
+         bone = uBones[int(aBoneID0)] * aBoneWeight0;
       }
       if (aBoneID1 != -1) {
-         bone += uBones[aBoneID1] * aBoneWeight1;
+         bone += uBones[int(aBoneID1)] * aBoneWeight1;
       }
       if (aBoneID2 != -1) {
-         bone += uBones[aBoneID2] * aBoneWeight2;
+         bone += uBones[int(aBoneID2)] * aBoneWeight2;
       }
       if (aBoneID3 != -1) {
-         bone += uBones[aBoneID3] * aBoneWeight3;
+         bone += uBones[int(aBoneID3)] * aBoneWeight3;
       }
       if (aBoneID4 != -1) {
-         bone += uBones[aBoneID4] * aBoneWeight4;
+         bone += uBones[int(aBoneID4)] * aBoneWeight4;
       }
    }
